Justly

Justly.Chord

Justly.edit_song

Justly.make_schedule

Justly.pedal

Justly.pluck

Justly.Chord — Type

mutable struct Chord

A Julia representation of a chord. Pass a vector of Chords to edit_song.

Justly.edit_song — Method

function edit_song(song; ramp = 0.1s, options...)

Use to edit songs interactively. The interface might be slow at first while Julia is compiling.

song is a YAML string or a vector of Chords.

ramp is the onset/offset time, in time units (like s).

options will be passed to make_schedule.

For more information, see the README.

Try running ENV["QT_QPA_PLATFORM"] = "xcb" on Wayland.

julia> using Justly

julia> song = Chord[];

julia> edit_song(song)

Justly.make_schedule — Method

function make_schedule(song;
    beat_duration = 0.6s,
    initial_key = 220Hz,
    make_envelope = pedal,
    sample_rate = 44100Hz,
    volume = 0.15,
    wave = SawTooth(7)
)

Create an AudioSchedule from your song.

song is a YAML string or a vector of Chords.

beat_duration is the duration of a beat, with time units (like s).

initial_key is initial key of your song, in frequency units (like Hz).

make_envelope is a function to make an envelope, like pluck or pedal.

sample_rate is the sample rate, in frequency units (like Hz).

volume, ranging from 0-1, is the volume that a single voice is played at.

wave is a function which takes an angle in radians and returns an amplitude between -1 and 1.

For more information, see the README.

For example, to create a simple I-IV-I figure,

julia> using Justly

julia> using YAML: load

julia> make_schedule("""
            - words: "I"
              notes:
                - {}
                - interval: "3/2"
                - interval: "5/4o1"     
            - words: "IV"
              interval: "2/3"
              notes:
                - interval: "3/2"
                - interval: "5/4o1"
                - interval: "o2"
            - words: "I"
              interval: "3/2"
              notes:
                - {}
                - interval: "3/2"
                - interval: "5/4o1"
              
         """)
1.85 s 44100.0 Hz AudioSchedule

Top-level lists will be unnested, so you can use YAML anchors to repeat themes.

julia> make_schedule("""
            - &fifth
                - notes:
                    - {}
                    - interval: "3/2"
            - *fifth
        """)
1.25 s 44100.0 Hz AudioSchedule

Justly.pedal — Method

pedal(duration; slope = 1 / 0.1s, peak = 1, overlap = 1/2)

You can use pedal to make an envelope with a sustain and ramps at the beginning and end. overlap is the proportion of the ramps that overlap.

Justly.pluck — Method

pluck(duration; decay = -2.5 / s, slope = 1 / 0.005s, peak = 1)

You can use pluck to make an envelope with an exponential decay and ramps at the beginning and end.