Justly

Justly.Chord

Justly.edit_song

Justly.make_schedule

Justly.pedal

Justly.pluck

Justly.Chord — Type

Chord(dictionary::Dict)

A Julia representation of a chord. Pass a vector of Chords to edit_song.

Justly.edit_song — Method

function edit_song(chords::Vector{Chord}; ramp = 0.1s, options...)

Open an interactive interface where you can interactively write Justly text. Once you have finished writing, you can copy the results to the clipboard as YAML. Then, you can use make_schedule to play your song. Might be slow at first while Julia is compiling.

chords should be a vector of [Chord]s to start editing. ramp is the time it takes for a note to switch from on to off, in time units (like s). options will be passed to make_schedule.

The first interval in the chord will modulate the key, and tells how many beats before the next key change. You can set beats to 0 to overlap, or to a negative number to "travel back in time". The rest of the intervals in the chord will play notes with a given duration. Their interval will show their relationship to the key. You can use words to as a way to keep track of your position in a song, or to make performance notes. For more information, see the README.

julia> using Justly

julia> chords = Chord[];

julia> edit_song(chords; test = true)

Justly.make_schedule — Method

function make_schedule(song;
    beat_duration = 0.6s,
    initial_key = 220Hz,
    make_envelope = pedal,
    sample_rate = 44100Hz,
    volume = 0.15,
    wave = SawTooth(7)
)

Play music in Justly notation.

song can be in YAML or a list of Chords.

beat_duration is duration of a beat with time units (like s).

initial_key is initial key of your song, in frequency units (like Hz).

make_envelope is should be a function to make an envelope, like pluck or pedal.

sample_rate is the sample rate in frequency units (like Hz).

volume will be the volume that a single voice is played at, ranging from 0-1.

wave should be a function which takes an angle in radians and returns an amplitude between 0 and 1.

For example, to create a simple I-IV-I figure,

julia> using Justly

julia> using YAML: load

julia> make_schedule(load("""
            - words: "I"
              notes:
                - {}
                - interval: "3/2"
                - interval: "5/4o1"     
            - words: "IV"
              interval: "2/3"
              notes:
                - interval: "3/2"
                - interval: "5/4o1"
                - interval: "1o2"
            - words: "I"
              interval: "3/2"
              notes:
                - {}
                - interval: "3/2"
                - interval: "5/4o1"
              
         """))
1.85 s 44100.0 Hz AudioSchedule

Note also that Justly will unnest top-level lists, so you use YAML anchors to repeat themes.

julia> make_schedule(load("""
            - &fifth
                - notes:
                    - {}
                    - interval: "3/2"
            - *fifth
        """))
1.25 s 44100.0 Hz AudioSchedule

Justly.pedal — Method

pedal(duration; slope = 1 / 0.1s, peak = 1, overlap = 1/2)

A sustain with steep ramps on either side. Overlap is the proportion of the ramps that overlap.

Justly.pluck — Method

pluck(duration; decay = -2.5 / s, slope = 1 / 0.005s, peak = 1)

The default envelope function used for make_schedule. An exponential decay with ramps on either side.