Justly

Justly.edit_song

Justly.pluck

Justly.read_justly

Justly.edit_song — Method

function edit_song(song_file; 
    test = false, 
    keywords...
)

Use to edit songs interactively. The interface might be slow at first while Julia is compiling.

song_file is a YAML string or a vector of Chords. Will be created if it doesn't exist.

keywords will be passed to read_justly.

For more information, see the README.

julia> using Justly

julia> edit_song(joinpath(pkgdir(Justly), "examples", "simple.justly"); test = true)

julia> edit_song(joinpath(pkgdir(Justly), "not_a_folder", "simple.justly"); test = true)
ERROR: ArgumentError: Folder doesn't exist!
[...]

Justly.pluck — Method

pluck(duration; ramp_duration = 0.05s, decay_rate = -4/s)

You can use pluck to make an envelope with an exponential decay_rate and ramps at the beginning and end.

ramp_duration is the duration of the ramps at the beginning and end.

decay_rate is the continuous negative decay rate.

Justly.read_justly — Method

function read_justly(song_file;
    volume = 0.2,
    frequency = 200.0,
    tempo = 200.0,
    wave = SawTooth(7),
    make_envelope = pluck
    
)

Create a Song from a song file.

volume is the starting volume of a single voice, between 0 and 1. Use this if the user doesn't specify a starting volume in the song file. To avoid peaking, lower the volume for songs with many voices.

frequency is the starting frequency, in Hz. Use this if the user doesn't specify a starting frequency in the song file.

tempo is the tempo of the song, in beats per minute. This is the tempo of 1 indivisible beat, so for songs which subdivide notes, you will need to multiply the tempo accordingly.

wave is a function which takes an angle in radians and returns an amplitude between -1 and 1. Use this to change the timbre of the sound. The default is SawTooth(7), a saw-tooth wave with the first 7 fundamental frequencies.

make_envelope is a function which takes a duration in seconds and returns an envelope. The default is pluck.

julia> using Justly

julia> song = read_justly(joinpath(pkgdir(Justly), "examples", "simple.justly"));

julia> print(song)
Frequency: 220.0 Hz
Volume: 0.2
Tempo: 200.0 bpm
# I
1: 1, 5/4 at 2.0, 3/2
# IV
2/3: 3/2, 1o1 at 2.0, 5/4o1
# I
3/2 for 2: 1 for 2, 5/4 for 2 at 2.0, 3/2 for 2

You can add a song to an AudioSchedule

julia> using AudioSchedules: AudioSchedule

julia> audio_schedule = AudioSchedule()
0.0 s 44100.0 Hz AudioSchedule

julia> push!(audio_schedule, song)

julia> audio_schedule
1.27 s 44100.0 Hz AudioSchedule

You can convert a song directly to a SampleBuf, with a default sample rate of 44100Hz.

julia> using SampledSignals: SampleBuf

julia> SampleBuf(song)
56007-frame, 1-channel SampleBuf{Float64, 1}
1.27s sampled at 44100.0Hz
▅▆▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▆▆▆▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▆▆▆▆▇▇▇▇▇▇▇▇▇▇▇▇▇▇▆▆▆▆▆▆▆▆▆▅▆▆▆▆▆▆▅▅▅▅▅▅▅▄▄▃