Justly

Justly.edit_song

Justly.make_music_xml

Justly.pedal

Justly.read_justly

Justly.edit_song — Method

function edit_song(song_file; 
    ramp = 0.1s, 
    number_of_tasks = nthreads() - 2, 
    test = false, 
    keyword_arguments...
)

Use to edit songs interactively. The interface might be slow at first while Julia is compiling.

song_file is a YAML string or a vector of Chords. Will be created if it doesn't exist.

number_of_tasks is the number of tasks to use to process data. Defaults to 2 less than the number of threads; we need 1 master thread for QML and 1 master thread for AudioSchedules.

If test is true, will open the editor briefly to test it.

keyword_arguments will be passed to read_justly.

For more information, see the README.

julia> using Justly

julia> edit_song(joinpath(pkgdir(Justly), "test", "song.justly"); test = true)

Justly.make_music_xml — Method

make_music_xml(
    song;
    key_fifths = 0,
    beats_per_quarter_note = 1,
    time_signature_numerator = 4,
    time_signature_denominator = 4,
)

Make a music xml file from a song. You can save it to a new file with LightXML.save_file.

Key fifths is the number of sharps and flats in the key signature. So if it's +1, the key is G major/E minor, and if it's -1, the key is F major/D minor.

julia> using Justly

julia> using LightXML: save_file

julia> cd(joinpath(pkgdir(Justly), "test"))

julia> song = read_justly("song.justly");

julia> xml_song = make_music_xml(song, beats_per_quarter_note = 2);

julia> save_file(xml_song, "test_song.musicxml");

julia> read("test_song.musicxml", String) == read("song.musicxml", String)
true

julia> rm("test_song.musicxml")

Justly.pedal — Method

pedal(duration; slope = 1 / 0.1s, peak = 1, overlap = 1/2)

You can use pedal to make an envelope with a sustain and ramps at the beginning and end. overlap is the proportion of the ramps that overlap.

Justly.read_justly — Method

function read_justly(song_file;
    wave = SawTooth(7),
    make_envelope = pedal,
    ramp = 0.1s,
    sample_rate = 44100.0Hz,
    volume = 0.1
)

Create a Song from a song file.

make_envelope is a function to make an envelope, like pedal.

ramp is the onset/offset time, in time units (like s).

sample_rate is the sample rate, in frequency units (like Hz).

volume, ranging from 0-1, is the volume that a single voice is played at.

wave is a function which takes an angle in radians and returns an amplitude between -1 and 1.

julia> using Justly

julia> song = read_justly(joinpath(pkgdir(Justly), "test", "song.justly"));

julia> print(song)
220.0 Hz; 800.0 bpm
# first chord
1 for 19: 1 for 3, 5/4 for 5, 3/2 for 8, 5/4o1 for 19
1 for -1: 1 for 1

You can create an AudioSchedule from a song.

julia> using AudioSchedules: AudioSchedule

julia> AudioSchedule(song)
1.55 s 44100.0 Hz AudioSchedule