Justly

Justly.Chord

Justly.edit_song

Justly.pedal

Justly.read_song

Justly.Chord — Type

mutable struct Chord

A Julia representation of a chord. Pass a vector of Chords to edit_song.

Justly.edit_song — Method

function edit_song(song_file; 
    ramp = 0.1s, 
    number_of_tasks = nthreads() - 2, 
    test = false, 
    keyword_arguments...
)

Use to edit songs interactively. The interface might be slow at first while Julia is compiling.

song_file is a YAML string or a vector of Chords. Will be created if it doesn't exist.

number_of_tasks is the number of tasks to use to process data. Defaults to 2 less than the number of threads; we need 1 master thread for QML and 1 master thread for AudioSchedules.

If test is true, will open the editor briefly to test it.

keyword_arguments will be passed to read_song.

For more information, see the README.

julia> using Justly

julia> edit_song(joinpath(pkgdir(Justly), "test", "test_song_file.yml"); test = true)

Justly.pedal — Method

pedal(duration; slope = 1 / 0.1s, peak = 1, overlap = 1/2)

You can use pedal to make an envelope with a sustain and ramps at the beginning and end. overlap is the proportion of the ramps that overlap.

Justly.read_song — Method

function read_song(song_file;
    wave = SawTooth(7),
    make_envelope = pedal,
    ramp = 0.1s,
    sample_rate = 44100.0Hz,
    volume = 0.1
)

Create an AudioSchedule from a song file.

make_envelope is a function to make an envelope, like pedal.

ramp is the onset/offset time, in time units (like s).

sample_rate is the sample rate, in frequency units (like Hz).

volume, ranging from 0-1, is the volume that a single voice is played at.

wave is a function which takes an angle in radians and returns an amplitude between -1 and 1.

julia> using Justly

julia> read_song(joinpath(pkgdir(Justly), "test", "test_song_file.yml"))
0.55 s 44100.0 Hz AudioSchedule