Justly

Justly.DEFAULT_INSTRUMENTS

Justly.Instrument

Justly.default_wave

Justly.edit_justly

Justly.pulse!

Justly.read_justly

Justly.sustain!

Justly.DEFAULT_INSTRUMENTS — Constant

DEFAULT_INSTRUMENTS

The default [Instrument]s available in read_justly and edit_justly, namely,

[Instrument(pulse!, "pulse!"), Instrument(sustain!, "sustain!")]

See pulse! and sustain!.

Justly.Instrument — Type

Instrument(note_function!, name)

Use an Instrument to play a note in the style of an instrument.

note_function! will be called as follows:

note_function!(audio_schedule, start_time, duration, volume, frequency)

where

audio_schedule is an AudioSchedules.AudioSchedule to add the note to.

start_time is a start time in seconds, like 0.0s.

duration is a duration in seconds, like 1.0s

volume is a ratio between 0 and 1.

frequency is the frequency of the note in hertz, like 440.0Hz.

Justly.default_wave — Method

default_wave(x)

A wave with many overtones, namely,

sin(x) +
0.2sin(2x) +
0.166sin(3x) +
0.133sin(4x) +
0.1sin(5x) +
0.066sin(6x) +
0.033sin(7x)

Justly.edit_justly — Function

function edit_justly(song_file, instruments = DEFAULT_INSTRUMENTS; 
    test = false
)

Use to edit songs interactively. The interface might be slow at first while Julia is compiling.

song_file is a YAML file. Will be created if it doesn't exist.

instruments are a vector of Instruments, with the default DEFAULT_INSTRUMENTS.

For more information, see the README.

julia> using Justly

julia> edit_justly(joinpath(pkgdir(Justly), "examples", "simple.yml"))

Justly.pulse! — Method

pulse!(audio_schedule, start_time, duration, volume, frequency;
    ramp_duration = 0.07s,
    decay_rate = -4/s
)

A function for an Instrument.

Uses the wave function default_wave. Add an envelope with an exponential decay_rate and ramps of ramp_duration at the beginning and end.

Justly.read_justly — Function

read_justly(file, instruments = DEFAULT_INSTRUMENTS)

Create a Song from a song file.

instruments are a vector of Instruments, with the default DEFAULT_INSTRUMENTS.

julia> using Justly

julia> song = read_justly(joinpath(pkgdir(Justly), "examples", "simple.yml"));

You can add a song to an AudioSchedule at a start time. You can use this to overlap multiple songs at differnt times.

julia> using AudioSchedules: AudioSchedule, Hz, s

julia> audio_schedule = AudioSchedule(sample_rate = 44100Hz)
0.0 s 44100.0 Hz AudioSchedule

julia> push!(audio_schedule, song, 0.0s)

julia> push!(audio_schedule, song, 1.0s)

julia> audio_schedule
2.27 s 44100.0 Hz AudioSchedule

Justly.sustain! — Method

sustain!(audio_schedule, start_time, duration, volume, frequency;
    ramp_duration = 0.07s
)

A function for an Instrument.

Uses the wave function default_wave. Add an envelope with ramps of ramp_duration at the beginning and end.